chore: extract time delta text
