+++ src/dotnet/Chat.UI.Blazor/Components/HistoricalPlaybackSubHeader/HistoricalPlaybackSubHeader.razor
@@ -48,7 +48,7 @@
                     @if (m.PlayingAt is { } playingAt) {
                         var localTime = DateTimeConverter.ToLocalTime(playingAt);
                         var absoluteTime = $"{localTime.ToLongDateString()} at {localTime.ToShortTimeString()}";
-                        var relativeTime = LiveTime.GetDeltaText(playingAt, Clocks.SystemClock.Now);
+                        var relativeTime = LiveTime.GetDeltaText(playingAt);
                         <span>
                             Playing audio recorded
                             <span class="c-moment" title="@absoluteTime">@relativeTime</span>
@@ -81,7 +81,6 @@
     private IChats Chats => Hub.Chats;
     private ChatUI ChatUI => Hub.ChatUI;
     private ChatPlayers ChatPlayers => Hub.ChatPlayers;
-    private MomentClockSet Clocks => Hub.Clocks();
     private DateTimeConverter DateTimeConverter => Hub.DateTimeConverter;
 
     [CascadingParameter] public ScreenSize ScreenSize { get; set; }
